factor out openssa.l2.util.misc.AskAnsPair & format_other_result(...)
