+++ openssa/l2/reasoning/base/__init__.py
@@ -9,9 +9,10 @@ from typing import TYPE_CHECKING
 from openssa.l2.reasoning.abstract import AbstractReasoner
 from openssa.l2.knowledge._prompts import knowledge_injection_lm_chat_msgs
 from openssa.l2.task.status import TaskStatus
+from openssa.l2.util.misc import format_other_result
 
 from ._prompts import (RESOURCE_QA_CONSO_PROMPT_TEMPLATE, RESOURCE_QA_AND_OTHER_RESULTS_CONSO_PROMPT_TEMPLATE,
-                       OTHER_RESULTS_CONSO_PROMPT_TEMPLATE, format_other_result)
+                       OTHER_RESULTS_CONSO_PROMPT_TEMPLATE)
 
 if TYPE_CHECKING:
     from openssa.l2.planning.abstract.plan import AskAnsPair
